refactor: Split script into multiple modules
