+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="simpler-grants-sandbox",
+    version="0.1.0",
+    description="Simpler Grants Sandbox",
+    packages=find_packages(),
+    python_requires=">=3.8",
+)
